Allow keeping the K^th weight fixed during optimization
Note that keeping _all_ weights fixed was already possible.

This commit adresses the case where the K^th weight is fixed, but at least
two[*] other weights are not. This is a bit complicated because the K^th weight
is only given implicitly in the flattened DesignMeasure representation. The
solution is to only move parallel to the diagonal face of the (solid) simplex.

An alternative would be to permute the designpoints such that a non-fixed one
is in the final position. In trying this out the resulting code seemed much
less clear, especially as it would return a DesignMeasure and optimizer states
where the fixed elements are in other positions than the user specified in the
function call. One could retroactively permute also them, but this would make
the code even less clear.

[*]: Due to the sum constraint, fixing all weights but one is equivalent to
     fixing all of them.