Remove information matrix regularization feature
It results in too many problems with setting up a logically coherent
design problem and with interpreting the results. At the moment it is
not clear whether there is a well-defined asymptotic derivation of the
objective function that results from using regularization.

The abstract supertype for normal approximations is kept for now,
although it has only one subtype.