Add graceful (and correct) handling of singular designs
Exceptions are now caught. Objective functions return -Inf, Gateaux
derivatives return NaN.

Also fix a subtle bug in case the untransformed information matrix is
non-singular, but the transformed one is. The latter it is not
inverted when evalutaing the objective, thus it throws no exception.
The objective function, however, erroneously returned Inf. This in
turn confused the particle swarm optimizer.