Add explicit type for difference between AbstractPoints
The PSO implementation now expects the difference between two
`AbstractPoint`s to be an `AbstractPointDifference` instead of a
`Vector{<:Real}`. For `DesignMeasure`s, these differences are now
represented as `SignedMeasure`s.

This change greatly increases the readability of the PSO methods for
`DesignMeasure` because we can now exploit the added structure in the
difference objects. We don't have to do the index computations into a
difference vector any more.

Additionally, all methods dealing with `AbstractPoint`s are now
prefixed with `ap_`.

Note that we could _not_ have used `DesignMeasure`s themselves for
representing the differences because their weights must be
non-negative and sum to one.