Implement the strategy pattern for refinement
The functionality of former method `refine_design` is now contained in
the `Exchange` strategy for `solve`. The new return value is an
`ExchangeResult` that wraps vectors of `OptimizationResult`s from the
direction-finding and reweighting steps.

Along the way this fixes a bug in the getting started vignette where
`refine_design` was called for the wrong design problem.