Remove as_matrix and corresponding constructor
The matrix representation seems not to be unique in the
literature (e.g. Fedorov/Leonov have weights in the _last_ row and not
in the first). So we remove it to keep the package more minimal. For
users, it is trivial to create a matrix from a `DesignMeasure` with
`[reduce(hcat, points(d)); weights(d)']`. Since the inner constructor
now takes a matrix representation of the design points, the converse
is also simple.