Change simplify to drop after merge
This prevents points that are close, but individually too light, from
being dropped before they can be merged. Especially with the Exchange
algorithm there are plausible situations where this would generate
spuriously singular designs.

This change is technically breaking, but the differences in the
vignette examples are only on the magnitude of numerical inaccuracy.