Refactor model-specific code out of `informationmatrix!`
Now, `informationmatrix` only dispatches on the subtype of
`NormalApproximation`, but is generic for the model type.

The averaged Fisher information matrix is now computed specifically
for the abstract model type. Currently, there is an implementation
`average_fishermatrix!` for `M <: NonlinearRegression`.

This change has the nice effect that the signatures of the following
functions are now compatible with any `M <: Model`:

- `objective!`
- `gateauxderivative!`
- `inverse_information_matrices`
- `plot_gateauxderivative`