Don't use floating point exponentiation in dtr example
Between julia v1.9.4 and v1.10.0 the last bit of the result of
floating point exponetiation changed. For example, look at the
bitstring of `0.1^5.0`. This lead to noticeably different results in
the log-equidistant dtr design example. As a workaround that produces
identical results under both versions we avoid exponentiation for now.

This changes the value of the objective function slightly for problem
dp5, and by implication also the efficiencies for s5 are a bit
different now.